changes from pr. need to get working now
